requirements.txt
h5py==2.10.0
joblib==0.14.1
matplotlib==2.2.2
numpy==1.17.0
pandas==0.25.1
requests==2.21.0
scikit-learn==0.21.2
scipy==1.4.1
seaborn==0.8.1
tensorboard==2.1.1
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tqdm==4.33.0
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.1
zipp==3.1.0
flask
gevent